adapt zdx to use zng instead of ad hoc sorted tables
This commit adapts zar and zdx to use zng files as the underlying
representation of sorted index keys.  Now that we have stream
boundaries within zng files, we can use these boundaries to create
hierarchical indexes.

This commit also generalizes the zar command funcationality to handle
arbitrary indexing patterns of types or fields.  Once the zng indexing
machinery was in place, this change was relatively straightforward.